feat: handle invalid state transition refactor: move exception to its namespace
