+++ src/Client.php
@@ -12,6 +12,8 @@ use GuzzleHttp\MessageFormatter;
 use GuzzleHttp\Middleware;
 use GuzzleHttp\Psr7\Request;
 use JsonException;
+use Keboola\JobQueueInternalClient\Exception\ClientException;
+use Keboola\JobQueueInternalClient\Exception\StateTargetEqualsCurrentException;
 use Keboola\JobQueueInternalClient\JobFactory\Job;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -237,6 +239,16 @@ class Client
             $response = $this->guzzle->send($request);
             $data = json_decode($response->getBody()->getContents(), true, self::JSON_DEPTH, JSON_THROW_ON_ERROR);
             return $data ?: [];
+        } catch (\GuzzleHttp\Exception\ClientException $e) {
+            try {
+                $body = json_decode($e->getResponse()->getBody()->getContents(), true, 512, JSON_THROW_ON_ERROR);
+            } catch (Throwable $e2) {
+                throw new ClientException($e->getMessage(), $e->getCode(), $e2);
+            }
+            if (!empty($body['context']['stringCode']) && ($body['context']['stringCode'] === 'statusTargetEqualsCurrent')) {
+                throw new StateTargetEqualsCurrentException($e->getMessage(), $e->getCode(), $e);
+            }
+            throw new ClientException($e->getMessage(), $e->getCode(), $e);
         } catch (GuzzleException $e) {
             throw new ClientException($e->getMessage(), $e->getCode(), $e);
         } catch (JsonException $e) {
