+++ src/Exception/StateTargetEqualsCurrentException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\JobQueueInternalClient\Exception;
+
+class StateTargetEqualsCurrentException extends ClientException
+{
+}
